refactor: sbpagination
